# Optimize dynamic unmarshaling